requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
atlastk==0.13.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dictdiffer==0.9.0
exceptiongroup==1.2.0
fastapi==0.104.1
fastapi-debug-toolbar==0.5.0
fs-moysklad-api @ git+https://github.com/fogstream/fs-moysklad-api@667d05fcbbf574b46d46233e407b47cc926dcc5f
gunicorn==21.2.0
h11==0.14.0
idna==3.6
iso8601==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyinstrument==4.6.1
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
tomlkit==0.12.3
tortoise-orm==0.20.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
